+++ plugin/gui.py
@@ -1,12 +1,33 @@
-from aqt import mw
+from aqt import mw, gui_hooks
 from aqt.qt import *
 from aqt.utils import showInfo
 from aqt.operations import QueryOp
 
-from .db_utils import init_db, android_gen, get_num_files_per_source
+from .db_utils import (
+    init_db,
+    android_gen,
+    get_num_files_per_source,
+    table_exists_and_has_data,
+    table_must_be_updated,
+    update_db_version,
+)
 
 
-def regenerate_database_operation():
+def attempt_init_db_gui():
+    """
+    attempts to initialize the db
+    if not initialized, runs with gui instead of freezing Anki
+    """
+
+    if not table_exists_and_has_data():
+        regenerate_database_operation()
+    elif table_must_be_updated():
+        regenerate_database_operation("Updating local audio database.")
+
+
+def regenerate_database_operation(msg="Generating local audio database."):
+    update_db_version()
+
     op = QueryOp(
         # the active window (main window in this case)
         parent=mw,
@@ -20,7 +41,9 @@ def regenerate_database_operation():
 
     # if with_progress() is not called, no progress window will be shown.
     # note: QueryOp.with_progress() was broken until Anki 2.1.50
-    op.with_progress().run_in_background()
+    op.with_progress(
+        f"{msg}\nThis may take a while..."
+    ).run_in_background()
 
 
 def regenerate_database_action() -> int:
@@ -46,7 +69,9 @@ def generate_android_database_operation():
 
     # if with_progress() is not called, no progress window will be shown.
     # note: QueryOp.with_progress() was broken until Anki 2.1.50
-    op.with_progress().run_in_background()
+    op.with_progress(
+        "Generating local audio database (for Android).\nThis may take a while..."
+    ).run_in_background()
 
 
 def generate_android_database_action():
@@ -65,6 +90,9 @@ def action_get_num_files_per_source():
 
 
 def init_gui():
+    # must use a hook, so Anki can actually show the process window
+    gui_hooks.main_window_did_init.append(attempt_init_db_gui)
+
     menu_local_audio = mw.form.menuTools.addMenu("Local Audio Server")
 
     # regenerate regular database (entries.db)